requirements.txt
streamlit>=1.38.0
pandas
pillow
qrcode
reportlab
opencv-python-headless
numpy
streamlit-webrtc
